Working on webhook feature.

+++ sql/queries/chirps.sql
@@ -16,7 +16,7 @@ SELECT * FROM chirps WHERE user_id=$1;
 SELECT * FROM chirps WHERE id=$1;
 
 -- name: DeleteChirp :exec
-DELETE FROM chirps WHERE id=$1;
+DELETE FROM chirps WHERE id=$1 AND user_id=$2;
 
 -- name: GetAllChirps :many
 SELECT * FROM chirps
